Move the class into more generic namespace.
