+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Components/Menus/LockStateMenu.cs
@@ -4,6 +4,7 @@
 using System.Collections.Generic;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Rulesets.Karaoke.Edit.Components.Menus;
+using osu.Game.Rulesets.Karaoke.Graphics.UserInterface;
 using osu.Game.Rulesets.Karaoke.Objects.Types;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics;
 
@@ -12,9 +13,9 @@ namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Components.Menus;
 /// <summary>
 /// If click the lock icon in <see cref="LyricEditor"/>, will apply <see cref="LockState.Partial"/> or <see cref="LockState.Full"/>
 /// </summary>
-public class LockStateMenu : EnumMenu<LockState>
+public class LockStateMenuItem : BindableEnumMenuItem<LockState>
 {
-    public LockStateMenu(KaraokeRulesetLyricEditorConfigManager config, string text)
+    public LockStateMenuItem(KaraokeRulesetLyricEditorConfigManager config, string text)
         : base(config.GetBindable<LockState>(KaraokeRulesetLyricEditorSetting.ClickToLockLyricState), text)
     {
     }
