+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Components/Menus/LyricEditorPreferLayoutMenu.cs
@@ -3,13 +3,14 @@
 
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Rulesets.Karaoke.Edit.Components.Menus;
+using osu.Game.Rulesets.Karaoke.Graphics.UserInterface;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Components.Menus;
 
-public class LyricEditorPreferLayoutMenu : EnumMenu<LyricEditorLayout>
+public class LyricEditorPreferLayoutMenuItem : BindableEnumMenuItem<LyricEditorLayout>
 {
-    public LyricEditorPreferLayoutMenu(KaraokeRulesetLyricEditorConfigManager config, string text)
+    public LyricEditorPreferLayoutMenuItem(KaraokeRulesetLyricEditorConfigManager config, string text)
         : base(config.GetBindable<LyricEditorLayout>(KaraokeRulesetLyricEditorSetting.LyricEditorPreferLayout), text)
     {
     }
